保存

+++ si4463-stm32-库函数/Demo_SI4438_Stm32F103/Project_SI4438/Source_Files/drv_mcu/inc/drv_uart.h
@@ -39,7 +39,7 @@
 
 
 void drv_uart_init( uint32_t UartBaudRate );
-void drv_uart_tx_bytes( uint8_t* TxBuffer, uint8_t Length );
+void drv_uart_tx_bytes( uint8_t* TxBuffer, uint16_t Length );
 uint8_t drv_uart_rx_bytes( uint8_t* RxBuffer );
 extern uint32_t hex_to_dec(uint8_t *RxBuffer);
